// src/app/shared/models/mobile-grid-labels.ts
export class MobileGridLabels {
    isSortFilter?: Boolean;
    isStateFilter?: Boolean;
    isRegionFilter?: Boolean;
    isLocationFilter?: Boolean;
    isSpecialityFilter?: Boolean;
    isNextDateFilter?: Boolean;
    isEndDateFilter?: Boolean;
    isAppliedDate: Boolean;
    isDateRangeFilter?: Boolean;
    isStatusFilter?: Boolean;
    isCityFilter?: Boolean;
    isRecruiterFilter?: Boolean;
  }
  